empDaoImpl file 追加 empInfo service 部分修正（）

+++ src/main/java/com/egrina/planet/web/dao/emp/EmpWorkTimeDao.java
@@ -9,7 +9,7 @@ public interface EmpWorkTimeDao {
 
     public EmpWorkTime findById(Integer empWorkTimePk);
 
-    public void delete(long empWorkTimePk);
+    public void delete(Integer empWorkTimePk);
 
     public EmpWorkTime save(EmpWorkTime empWorkTime);
 }
